Comment middleman-specific autocmds
They are defined on a repo I will never again touch (and good riddance),
but they are useful settings to have around for reference.

This raises a major (okay, EXTREMELY trivial, but fundamental to this
trivial little dev subdomain) issue: how to generically identify a
Middleman app as opposed to a Rails or Sinatra one?

To be honest, I'm not sure it's even necessary to parse what sort of
ruby app it is: supposing this gets tweaked to, say, apply to any
project with a Gemfile in the root dir, will this interfere with
Rails.vim settings whatsoever?