Remove insert-mode nav shortcuts
It is super convenient and decisively faster to be able to just type
really rare but ergonomic af combinations like `kk` and `hh` or to just
drop the left pinky on control and `hjkl` around.

It is also a typing crutch that fundamentally undermines the logical
grouping of the whole diff from a single insert mode "session" to a
single, reusable (what's good, `:normal .`!) edit.