Remove ll and lll commands
I added a new `ll` script to the bin/ repo, which is sort of the best of
both worlds: it sets the env variable to pipe color escape sequences,
then sends the output of the `ls` command/options that had been the `ll`
alias to an awk script that added headers in a more readable,
sophisticated way than the old `lll` alias.

Given that color is preserved and the information better contextualized,
there was no longer any reason to prefer the old `ll` and make this new
script replace `lll` alone. So fuck it.