Reorganize plugins and `filetype plugin indent on`
The latter part feels like throwing shit at the wall; there was a
hideous syntax highlighting hiccup alongside my move to vim 8.0. Not
sure if I fixed it with this or if it was just some temporary weirdness,
but since it is, for the most part, working, I'm not fucking with my
panicked edits.

Goddamn it, I hate typing that, tho