Clean up aliases loading re: vim
N.B. "vim" refers to both vim and nvim

Vim shells out each :!command directly, without loading an interactive
shell. That is, vim sources `~/.zshenv`, but not `~/.zshrc`. This is an
issue re: aliases, because:

1) it is extremely nice and expected to have access to aliases when
   shelling out (noninteractively) from vim's (interactive) command line
2) it is poor form to load up the aliases in roughly every other
   possible noninteractive context, and besides, it makes emacs yell at
   me and that hurts my feelings

So with this patch, zshenv loads the aliases file if and only if it is
sourced from inside vim; and zshrc ALSO loads the aliases file, since
in the only scenario in which zshenv loads it, zshrc is not loaded.